modifying folder structure of monitoring.
