+++ k2s/test/e2e/addons/monitoring/02-securityenhanced/monitoring_security_test.go
